Set safecrlf option to true in git settings
From 'git help config'
CRLF conversion bears a slight chance of corrupting data. When it is enabled, Git will convert CRLF to LF during
commit and LF to CRLF during checkout. A file that contains a mixture of LF and CRLF before the commit cannot be
recreated by Git. For text files this is the right thing to do: it corrects line endings such that we have only LF
line endings in the repository. But for binary files that are accidentally classified as text the conversion can
corrupt data.